lib/github-actions.test.ts
import { assertEquals } from "https://deno.land/std@0.224.0/assert/assert_equals.ts";
import { assertThrows } from "https://deno.land/std@0.224.0/assert/assert_throws.ts";
import { GitHubActionsImpl } from "./github-actions.ts";
import { DetermineNextReleaseStepConfig } from "./steps/determine-next-release.ts";

Deno.test("getDetermineNextReleaseStepConfig returns undefined when env variable is not set", () => {
  Deno.env.delete("INPUT_ANALYZE_COMMITS_CONFIG");
  const githubActions = new GitHubActionsImpl();
  const result = githubActions.getDetermineNextReleaseStepConfig();
  assertEquals(result, undefined);
});

Deno.test("getDetermineNextReleaseStepConfig returns parsed config when env variable is set to valid JSON", () => {
  const expectedConfig: DetermineNextReleaseStepConfig = {
    branches: [
      {
        branch_name: "main",
        prerelease: false,
        version_suffix: "beta",
      },
    ],
  };

  const givenConfig = `
    {
      "branches": [
        {
          "branch_name": "main",
          "prerelease": false,
          "version_suffix": "beta"
        }
      ]
    }
  `
  Deno.env.set("INPUT_ANALYZE_COMMITS_CONFIG", givenConfig);
  assertEquals(new GitHubActionsImpl().getDetermineNextReleaseStepConfig(), expectedConfig)
});

Deno.test("getDetermineNextReleaseStepConfig throws error when env variable is set to invalid JSON", () => {
  Deno.env.set("INPUT_ANALYZE_COMMITS_CONFIG", "invalid-json");
  const githubActions = new GitHubActionsImpl();
  assertThrows(
    () => {
      githubActions.getDetermineNextReleaseStepConfig();
    },
    Error,
    "When trying to parse the GitHub Actions input value for analyze_commits_config, I encountered an error"
  );
});